refactor(desktop): share the theme shape check between registries
